requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
alembic==1.14.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
frozenlist==1.5.0
greenlet==3.1.1
idna==3.10
Mako==1.3.6
MarkupSafe==3.0.2
multidict==6.1.0
openai==0.28.0
propcache==0.2.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.36
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.17.2
